dispoapi/pyproject.toml
[project]
name = "dispoapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "openai>=1.30.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "modal>=0.73.0",
]
